+++ mayan/apps/document_parsing/tests/test_views.py
@@ -43,6 +43,13 @@ class DocumentContentViewTestMixin(object):
             }
         )
 
+    def _request_test_document_parsing_error_list_view(self):
+        return self.get(
+            viewname='document_parsing:document_parsing_error_list', kwargs={
+                'pk': self.test_document.pk,
+            }
+        )
+
 
 @override_settings(DOCUMENT_PARSING_AUTO_PARSING=True)
 class DocumentContentViewsTestCase(
@@ -97,7 +104,7 @@ class DocumentContentViewsTestCase(
 
     def test_document_page_content_view_with_access(self):
         self.grant_access(
-            permission=permission_content_view, obj=self.test_document
+            obj=self.test_document, permission=permission_content_view
         )
 
         response = self._request_test_document_page_content_view()
@@ -109,10 +116,12 @@ class DocumentContentViewsTestCase(
         response = self._request_test_document_content_download_view()
         self.assertEqual(response.status_code, 403)
 
-    def test_download_view_with_access(self):
-        self.expected_content_types = ('application/octet-stream; charset=utf-8',)
+    def test_document_parsing_download_view_with_access(self):
+        self.expected_content_types = (
+            'application/octet-stream; charset=utf-8',
+        )
         self.grant_access(
-            permission=permission_content_view, obj=self.test_document
+            obj=self.test_document, permission=permission_content_view
         )
 
         response = self._request_test_document_content_download_view()
@@ -123,6 +132,18 @@ class DocumentContentViewsTestCase(
                 ''.join(get_document_content(document=self.test_document))
             ),
         )
+
+    def test_document_parsing_error_list_view_no_permission(self):
+        response = self._request_test_document_parsing_error_list_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_parsing_error_list_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_parse_document
+        )
+
+        response = self._request_test_document_parsing_error_list_view()
+        self.assertEqual(response.status_code, 200)
 
 
 class DocumentTypeContentViewsTestMixin(object):
@@ -144,19 +165,17 @@ class DocumentTypeContentViewsTestCase(
 
     def test_document_type_parsing_settings_view_with_access(self):
         self.grant_access(
-            permission=permission_document_type_parsing_setup,
-            obj=self.test_document_type
+            obj=self.test_document_type,
+            permission=permission_document_type_parsing_setup
         )
 
         response = self._request_test_document_type_parsing_settings()
         self.assertEqual(response.status_code, 200)
 
 
-class DocumentContentToolsViewsTestCase(GenericDocumentViewTestCase):
-    _skip_file_descriptor_test = True
-
-    # Ensure we use a PDF file
-    test_document_filename = TEST_HYBRID_DOCUMENT
+class DocumentContentToolsViewsTestMixin(object):
+    def _request_document_parsing_error_list_view(self):
+        return self.get(viewname='document_parsing:error_list')
 
     def _request_document_parsing_tool_view(self):
         return self.post(
@@ -165,10 +184,33 @@ class DocumentContentToolsViewsTestCase(GenericDocumentViewTestCase):
             }
         )
 
+
+class DocumentContentToolsViewsTestCase(
+    DocumentContentToolsViewsTestMixin, GenericDocumentViewTestCase
+):
+    _skip_file_descriptor_test = True
+    auto_upload_document = False
+
+    # Ensure we use a PDF file
+    test_document_filename = TEST_HYBRID_DOCUMENT
+
     def _get_document_content(self):
-        return ''.join(list(get_document_content(document=self.test_document)))
+        return ''.join(
+            list(get_document_content(document=self.test_document))
+        )
+
+    def test_document_parsing_error_list_view_no_permission(self):
+        response = self._request_document_parsing_error_list_view()
+        self.assertEqual(response.status_code, 403)
+
+    def test_document_parsing_error_list_view_with_permission(self):
+        self.grant_permission(permission=permission_parse_document)
+
+        response = self._request_document_parsing_error_list_view()
+        self.assertEqual(response.status_code, 200)
 
     def test_document_parsing_tool_view_no_permission(self):
+        self.upload_document()
 
         response = self._request_document_parsing_tool_view()
         self.assertEqual(response.status_code, 200)
@@ -183,6 +225,8 @@ class DocumentContentToolsViewsTestCase(GenericDocumentViewTestCase):
 
     def test_document_parsing_tool_view_with_permission(self):
         self.grant_permission(permission=permission_parse_document)
+
+        self.upload_document()
 
         response = self._request_document_parsing_tool_view()
         self.assertEqual(response.status_code, 302)